bulk_file_organizer/requirements.txt
# Bulk File Organizer dependencies

pyyaml==6.0.2       # YAML parser for configuration files
pydantic==2.9.1     # Validation of configuration structures
pytest==8.3.2       # Unit testing framework
